rename component part 1
