+++ src/app/nip76/nip76-main/nip76-settings.component.spec.ts
